+++ lib/pages/form.ex
@@ -161,4 +161,18 @@ defmodule Pages.Form do
       do: List.keyreplace(attrs, "value", 0, {"value", value}),
       else: List.keydelete(attrs, "value", 0)
   end
+
+  defp update_value(attrs, "checkbox", inputs) do
+    {"name", name} = List.keyfind(attrs, "name", 0) || {"name", nil}
+    {"value", field_value} = List.keyfind(attrs, "value", 0)
+
+    checked? = List.keymember?(attrs, "checked", 0)
+    value = Map.get(inputs, name) |> to_string()
+
+    cond do
+      checked? && field_value == value -> attrs
+      field_value == value -> [{"checked", "checked"} | attrs]
+      true -> List.keydelete(attrs, "checked", 0)
+    end
+  end
 end